refactor: 네이밍 변경 및 오버로딩 메서드 추가
