+++ OpenMarket/OpenMarket/Networking/Protocol/URLSessionProtocol.swift
@@ -10,11 +10,19 @@ import Foundation
 protocol URLSessionProtocol {
     func dataTask(with url: URL,
                   completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol
+    
+    func dataTask(with url: URLRequest,
+                  completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol
 }
 
 extension URLSession: URLSessionProtocol {
     func dataTask(with url: URL,
                   completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
         return dataTask(with: url, completionHandler: completion)
+    }
+    
+    func dataTask(with urlRequest: URLRequest,
+                  completion: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
+        return dataTask(with: urlRequest, completionHandler: completion)
     }
 }
